Django/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1,<3.0
django-cors-headers>=3.10,<4.0
python-dotenv>=0.19,<1.0
django-filter